requirements.txt
Flask==3.1.2
gunicorn
joblib==1.4.2
numpy==2.3.3
pandas==2.3.2
python-dotenv==1.1.1
pytz==2024.1
requests==2.32.5
schedule==1.2.2
scikit-learn==1.7.2
tqdm==4.67.1
xgboost==3.0.5
